refactor: Refactor ResultItem
