+++ src/components/home/Result/ResultItem.tsx
@@ -9,5 +9,9 @@ type Props = {
 export function ResultItem({ type }: Props) {
   const option = ITEM_TYPE[type];
 
-  return <>{option.expanded ? <ResultEntity option={option} /> : <ResultList option={option} />}</>;
+  return (
+    <>
+      {option.expanded ? <ResultEntity optionType={option} /> : <ResultList optionType={option} />}
+    </>
+  );
 }
